requirements.txt
# Google Gen AI SDK (Gemini 3) — needs 1.51.0+
google-genai>=1.51.0

# YouTube Data API + downloader
google-api-python-client
yt-dlp

# Headless API
fastapi
uvicorn
pydantic

# [Phase 1] YouTube Analytics API (OAuth). Uncomment when analytics.py lands.
# google-auth
# google-auth-oauthlib
# google-auth-httplib2
